# user-service/swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
components:
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
...
paths:
...
  "/reservations":
    post:
      summary: Create Reservation
      tags:
      - Reservations
      security:
      - bearer_auth: []
      parameters: []
      responses:
        '201':
          description: Tournament Created
          content: {}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                recipient:
                  type: integer
                start_timestamp:
                  type: integer
                  description: UNIX EPOCH date time
                end_timestamp:
                  type: integer
                  description: UNIX EPOCH date time
                kind:
                  type: string
                recipient_rsvp:
                  type: string
                note:
                  type: string
    get:
      summary: Search Reservations
      tags:
      - Reservations
      security:
      - bearer_auth: []
      parameters:
      - name: host
        in: query
        schema:
          type: integer
      - name: recipient
        in: query
        schema:
          type: integer
      - name: recipient_rsvp
        in: query
        schema:
          type: string
      - name: start_timestamp
        in: query
        schema:
          type: integer
      responses:
        '200':
          description: Search
          content: {}
...
  "/tournaments":
    post:
      summary: Creates New Tournament
      tags:
      - Tournaments
      security:
      - bearer_auth: []
      responses:
        '201':
          description: Tournament Created
          content: {}
    get:
      summary: Index Tournaments
      tags:
      - Tournaments
      security:
      - bearer_auth: []
      responses:
        '200':
          description: Get list of all tournaments
          content: {}
...
servers:
- url: http://{defaultHost}
  variables:
    defaultHost:
      default: localhost:3000
- url: https://{defaultHost}
  variables:
    defaultHost:
      default: localhost:3000
- url: https://nameless-spire-32644.herokuapp.com/